Skip multi-csvdiff file test
This test now needs to be broken up into multiple tests in order for the new
csvdiff executable to function properly (like a multi-exodiff test would).

The current test uses features not possible to exist in a non-stateful executable
(if a field exists in any csv file tested... then ignore missing fields in others).

Closes #12363